fix(#3921): create common fixture and conftest
